refactor(工厂模式): 代码位置调整
简单工厂和工厂模式代码位置调整
