+++ src/main/java/com/jt/test/demo1/factory/LoginStrategyFactory2.java
@@ -1,6 +1,6 @@
 package com.jt.test.demo1.factory;
 
-import com.jt.test.demo1.service.LoginStrategy;
+import com.jt.test.demo1.service.factory.LoginStrategy;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
