+++ src/main/java/com/jt/test/demo1/helper/OrderHelper.java
@@ -2,14 +2,11 @@ package com.jt.test.demo1.helper;
 
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONObject;
-import com.alibaba.nacos.shaded.org.checkerframework.checker.units.qual.A;
-import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.jt.test.demo1.common.HttpResult;
 import com.jt.test.demo1.convert.OrderConvert;
-import com.jt.test.demo1.domain.AmericanoCoffee;
 import com.jt.test.demo1.domain.Coffee;
 import com.jt.test.demo1.domain.entity.MemberPrice;
 import com.jt.test.demo1.domain.entity.Order;
@@ -17,16 +14,14 @@ import com.jt.test.demo1.domain.bo.OrderBO;
 import com.jt.test.demo1.domain.dto.OrderDTO;
 import com.jt.test.demo1.domain.vo.OrderVO;
 
-import com.jt.test.demo1.service.CoffeAbstractFactory;
+import com.jt.test.demo1.service.factory.CoffeAbstractFactory;
 import com.jt.test.demo1.service.MemberPriceService;
 import com.jt.test.demo1.service.OrderService;
 import com.jt.test.demo1.service.RedisService;
-import com.jt.test.demo1.service.impl.AmericanoFactroy;
-import com.jt.test.demo1.service.impl.CoffeeStore;
-import com.jt.test.demo1.service.impl.LatteFactory;
-import lombok.Data;
+import com.jt.test.demo1.service.factory.impl.AmericanoFactroy;
+import com.jt.test.demo1.service.factory.impl.CoffeeStore;
+import com.jt.test.demo1.service.factory.impl.LatteFactory;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.redis.core.RedisTemplate;
