+++ src/main/java/com/jt/test/demo1/service/factory/CoffeAbstractFactory.java
@@ -1,4 +1,4 @@
-package com.jt.test.demo1.service;
+package com.jt.test.demo1.service.factory;
 
 import com.jt.test.demo1.domain.Coffee;
