+++ src/main/java/com/jt/test/demo1/service/factory/LoginStrategy.java
@@ -1,4 +1,4 @@
-package com.jt.test.demo1.service;
+package com.jt.test.demo1.service.factory;
 
 /**
  * LoginStrategy
