+++ src/main/java/com/jt/test/demo1/service/factory/impl/LatteFactory.java
@@ -1,8 +1,8 @@
-package com.jt.test.demo1.service.impl;
+package com.jt.test.demo1.service.factory.impl;
 
 import com.jt.test.demo1.domain.Coffee;
 import com.jt.test.demo1.domain.LatteCoffee;
-import com.jt.test.demo1.service.CoffeAbstractFactory;
+import com.jt.test.demo1.service.factory.CoffeAbstractFactory;
 
 /**
  * LatteFactory
